NP-44875 moved query to MyRegistrations component
